rename: `@Enum`을 infra.annotation에서 domain.common.annotation으로 이동
