+++ src/main/java/es/princip/getp/domain/people/command/presentation/dto/request/UpdatePeopleRequest.java
@@ -1,5 +1,6 @@
 package es.princip.getp.domain.people.command.presentation.dto.request;
 
+import es.princip.getp.domain.common.annotation.Enum;
 import es.princip.getp.domain.member.command.annotation.EmailPattern;
 import es.princip.getp.domain.member.command.annotation.PhoneNumberPattern;
 import es.princip.getp.domain.member.command.domain.model.Email;
@@ -7,7 +8,6 @@ import es.princip.getp.domain.member.command.domain.model.Nickname;
 import es.princip.getp.domain.member.command.domain.model.PhoneNumber;
 import es.princip.getp.domain.people.command.application.command.UpdatePeopleCommand;
 import es.princip.getp.domain.people.command.domain.PeopleType;
-import es.princip.getp.infra.annotation.Enum;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
